reestruturação da pasta de user/controllers e criação de updateUserName
