+++ backend/controllers/user/createUser.go
@@ -1,0 +1,55 @@
+package user
+
+import (
+	"github.com/gofiber/fiber/v2"
+)
+
+type CreateUserRequestBody struct {
+	Name     string `json:"name" validate:"required"`
+	Email    string `json:"email" validate:"required"`
+	Password string `json:"password" validate:"required"`
+}
+
+func (handler *BaseHandler) CreateUser(context *fiber.Ctx) error {
+	body := new(CreateUserRequestBody)
+
+	if parseError := context.BodyParser(body); parseError != nil {
+		return context.Status(fiber.StatusBadRequest).JSON(fiber.Map{
+			"message": "Invalid request body",
+			"error":   parseError.Error(),
+		})
+	}
+
+	validationErrors := handler.Validator.Validate(body)
+
+	if validationErrors != nil {
+		errorMessages := make([]string, len(validationErrors))
+
+		for index := range validationErrors {
+			errorMessages[index] = validationErrors[index].Message
+		}
+
+		return context.Status(fiber.StatusBadRequest).JSON(fiber.Map{
+			"message": "Invalid request body",
+			"errors":  errorMessages,
+		})
+	}
+
+	newUser, serviceError := handler.Service.NewUser(
+		body.Name,
+		body.Email,
+		body.Password,
+	)
+
+	if serviceError != nil {
+		return context.Status(fiber.StatusBadRequest).JSON(fiber.Map{
+			"message": "user creation error",
+			"error":   serviceError.Error(),
+		})
+	}
+
+	return context.Status(fiber.StatusCreated).JSON(fiber.Map{
+		"message": "user created successfully",
+		"user":    newUser,
+	})
+}
